refactor(model-gen): tidy up and fix downloading of individual datasets
Original message:
"""
refactor(model-gen): tidy up and fix downloading of individual datasets

Most changes are documented in the notebook itself.

Key changes:
- General tidying up and added some documentation.
- Removed `dotenv` dependency. Instead, the Roboflow API key will be
  read from the `ROBOFLOW_API_KEY` environment variable. If not found,
  then you will be prompted.
- Fixed downloading and exporting of the COCO dataset, at least partially.
  For now, we still have to workaround a bug with FiftyOne — more
  details are available in comments in the notebook.

We will have to revisit the downloading of the COCO dataset since the
current workaround is not ideal. Many images we don't need are still
downloaded since filtering to "person" images is only done after the
download. This also means that we have a much smaller number of samples
than `max_samples` specified. There is a PR upstream that fixes the bug,
so we'll probably have to wait or specify the PR as a dependency. We'll
see.

Note that nothing after "Unzip dataset zip file (Not used anymore)" was
touched.
"""