feat(sentinel-server): start with FastAPI and create default user
The server now starts using FastAPI.

When the server is started, it will be check if the default user has
been created. If not, it will create one and, by default, save the
credentials to an sqlite database located under the user's data
directory. The default username is "root" and the default password
is "password".