refactor(server/ui): split UI code into modules
All UI code has now been moved to the `sentinel_server.ui` module. Under
this module, there is a module for each page (e.g., cameras, alerts,
dashboard, etc.). Each page module has its own router, which is added to
the application in the `setup()` function in `start.py`. These routers
are global. Additionally, to have the plugin manager accessible in other
modules, it needs to be global, so a `sentinel_server.globals` module
has been created to contain it.

`pages_shared()` has been moved to `sentinel_server/ui/__init__.py`. Not
sure if that's the best location, but it'll work for now.