fix(server/ui): fix camera table trying to refresh before video source manager has been initialised or loaded from the database
This shouldn't occur in a deployment setting, but, during development,
due to hot-reloading, it is possible for the table to try to refresh its
cells before the video source manager has been initialised or before it
has loaded video sources from the database. This would cause an
error to be raised.

This commit fixes that by making the table wait for the appropriate
async events introduced in 7bc7b870f04404dfbdb23be3305f1ed332d47529.