feat(server): allow attaching a detector to video source
Big change.

As with the video streams, detectors now have two variants:
async and sync.

`VideoSource`s now store detector information, similar to the ones
stored for video streams. In fact, the handling of detectors is
extremely similar to how video streams are handled. There is currently a
ton of code duplication (we can clean that up later).

The UI for the camera table has been updated to accommodate these
changes and now also shows the detector plugin and component. Checking
the "Enabled" box will also turn on the detector and attach it to the
video stream to make predictions. However, there is currently no way of
viewing the predictions from the UI.