feat(server): generate alerts from video sources
The video source manager will now create an alert emitter for each
video source when the video source is enabled. This emitter is
registered with the alert manager, which is responsible for managing
alert subscribers and emitters.

Currently, a placeholder `DatabaseSubscriber` is used to test that the
alerts are actually being generated. This subscriber currently just
prints the alerts to the server logs but will be refactored later to
actually store alerts in the database. However, before that, we should
implement filtering of frames to prevent spamming detections and alerts.