refactor(server): add interval parameter to `ReactiveDetector`
The interval represents the number of seconds to wait before passing the
frame to the detector. By default, the interval is 1 second, so
the underlying detector is only called every second. If the interval has
not yet passed frames are just forwarded without any detections. This
should hopefully massively improve performance.
.