refactor(model): add knife dataset
Added knife dataset for yolov8 model. There are a bunch of testing code
that should be removed in the final submittion. Have not tested the model but
the evalution function seems quite promising so I think can use this dataset.

For ray tune issue with local gpu, the work around will be tested another time
not too sure if it will be possible but worth a shot.