refactor(server): more robust stopping of video sources
Many async libraries seem to suppress the `TaskCancelled` exception,
which we previously relied on to stop video sources. This led to video
sources not stopping when disabled. This commit changes the stopping
mechanism to directly call `stop` on the underlying video stream so that
stopping video sources is more reliable.

Additional changes:
- Add callback mechanism for the video source errors.
- Update camera status in the camera table.