feat(server): implement subscriber management
Implements a subscriber manager to manage subscribers. The subscriber
manager is the "public" backend interface for the UI to register and
deregister subscribers. Behind the scenes, it talks to the alert
manager, which contains lower-level functionality to manage subscribers
and emitters.

Currently, loading from the database has not yet been implemented!