fix(server/ui): fix showing of alerts for deleted sources
The alert table filters alerts using the video source's name.
If there was an old camera that got deleted and had alerts,
and the user recreates a new camera using the same name,
alerts from the old camera would incorrectly show for the new camera.

This commit fixes this bug by adding an additional attribute to the
alerts stored in the database. This attribute is a boolean that
indicates whether the source has been deleted. If it has, then we know
that the alert should not be showed in the camera's alert table. In the
alerts page, alerts whose source has been deleted will still show, but
with `(deleted)` marker after the source name.